uniformfloat: switch scalar from double to float
This is going to simplify memory handling. Double precision is required
by the animation, but the uniform itself is a direct mapping to GPU
memory (where it's a float).

+++ libnodegl/nodes.h
@@ -275,7 +275,7 @@ struct uniform_priv {
         int i;
     } opt;
 
-    double scalar;
+    float scalar;
     float vector[4];
     float matrix[4*4];
     int ival;
